Directly override the classpath for javadoc.
With the value of javadoc-gen: classpath in pants.ini

In practice, pants was generating a javadoc command with an empty
classpath, and I couldn't understand (from their source) where exactly
it was _trying_ to get the classpath.  So rather than putting our
classpath where they need it, I just changed where its looking.

[QA]: None